+++ lib/src/visitor/async_evaluate.dart
@@ -854,10 +854,18 @@ final class _EvaluateVisitor
   }) async {
     var url = stylesheet.span.sourceUrl;
 
+    var currentConfiguration = configuration ?? _configuration;
     if (_modules[url] case var alreadyLoaded?) {
-      var currentConfiguration = configuration ?? _configuration;
       if (!_moduleConfigurations[url]!.sameOriginal(currentConfiguration) &&
-          currentConfiguration is ExplicitConfiguration) {
+          currentConfiguration is ExplicitConfiguration &&
+          // Don't throw an error if the module being loaded doesn't expose any
+          // configurable variables that could have been affected by the
+          // configuration in the first place. If the configuration defines a
+          // value that's not in the module, it'll still throw an error, but
+          // this avoids throwing confusing errors for `@forward`ed modules
+          // without configuration.
+          alreadyLoaded.couldHaveBeenConfigured(
+              MapKeySet(currentConfiguration.values))) {
         var message = namesInErrors
             ? "${p.prettyUri(url)} was already loaded, so it can't be "
                 "configured using \"with\"."
@@ -946,7 +954,7 @@ final class _EvaluateVisitor
     );
     if (url != null) {
       _modules[url] = module;
-      _moduleConfigurations[url] = _configuration;
+      _moduleConfigurations[url] = currentConfiguration;
       if (nodeWithSpan != null) _moduleNodes[url] = nodeWithSpan;
     }
 
@@ -2599,6 +2607,7 @@ final class _EvaluateVisitor
   Future<Value?> visitVariableDeclaration(VariableDeclaration node) async {
     if (node.isGuarded) {
       if (node.namespace == null && _environment.atRoot) {
+        _environment.markVariableConfigurable(node.name);
         if (_configuration.remove(node.name) case var override?
             when override.value != sassNull) {
           _addExceptionSpan(node, () {