+++ pkg/sass_api/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 15.8.0
+
+* No user-visible changes.
+
 ## 15.7.1
 
 * No user-visible changes.